introducao-ao-teste/src/main/java/edu/testesoftware/Mockito/AuthenticationService.java
package edu.testesoftware.Mockito;

//Faz a autenticação do usuário com base nas credenciais.
public class AuthenticationService {
    private UserRepository userRepository;

    public AuthenticationService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public boolean login(String username, String password) {
        User user = userRepository.findByUsername(username);
        if (user == null) {
            return false;
        }
        return user.getPassword().equals(password);
    }
}